Put each function in its own section
Data segments are also splited in their section.
It is one more step toward a link time dead code elimination.
Unused sections can be removed with "-cclib '-Wl,--gc-sections'" and the
result evaluated with "-cclib '-Wl,--print-gc-sections'".

This is just a hackish preliminary version, it assumes that the order of
sections is not modified by the linker. This may not be true. Later this
will be enforced by a linker script like:

SECTIONS
{
  .text     :
  {
    camlMod__code_begin = .;
    *(.text.caml__Mod.*)
    camlMod__code_end = .;
  }
  .data     :
  {
    camlMod__data_begin = .;
    *(.data.caml__Mod.*)
    camlMod__data_end = .;
  }
}