Inline_and_simplify check for validity of aliases
This is not useful currently, as an alias will be valid if and only if
the environment contain some information about the value. This will not
be the case anymore when approximation will be able to go back to an
environment similar to the one where they were created. This kind of
situation will occur when recursive functions will be able to return an
approximation even when not inlined.