Rework of meet functions:
- Introduce a module Meet_result, which allows tracking whether the result
the same as one of the arguments or actually more precise.
This is used to guarantee that infinite recursive meets are impossible.

- Make the functions add_equation and add_env_extension from Tying_env
systematically meet the equations. This required patching the join
algorithm in Typing_env_level to use these function correctly.

- Error when bottom equations are added to the environment.
The environment is equivalent to bottom whenever a single of its equations
is bottom, so no bottom equations should be added to it.
For equations coming from meet, the meet itself can return Bottom. For
other equations, the caller must correctly handle the bottom cases before
adding the equation.