[bugfix] Correctly propagate poison unboxed args
In some cases, a generated poison values for an unboxed arg was later
"transformed" into an available simple pointing to the "invalid"
constant for the adequate kind. This could bring problem where a
constant value is then passed to a deeper decisions that expects
a variant with no constant constructors.